Refactored invoice viewset
